feat: "cp -a" replace "cp -pr"
The -a option means -R and -p, plus a few other preservation options. It attempts to make a copy that's as close to the original as possible: same directory tree, same file types, same contents, same metadata (times, permissions, extended attributes, etc.). -r is identical with -R on Linux, it differs in some edge cases on some other unix variants.